delete route for campground can not be called unless user is logged in

+++ pages/api/campground/[id].js
@@ -57,6 +57,12 @@ handler
 			query: { id },
 		} = req;
 
+		const campground = await Campground.findById(id);
+
+		if (!campground.author.equals(id)) {
+			return res.redirect(`/campground/${id}`);
+		}
+
 		await Campground.findByIdAndDelete(id);
 	});
